Requirements met
The acceptance criteria is met in this version, however, it needs improvement.

The entire discount application process should be more like that of placing an order, where you will post and then response with a get. That'll take a lot of work out of the front end (the validation is awful right now).

Also generally make sure it adheres to MVC and the other criteria, check for useless methods, fields, etc.